requirements.txt
click==7.1.2
docopt==0.6.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
stomp.py==6.1.0
Werkzeug==1.0.1
